+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.Commitments.Shared/Services/CommitmentsService.cs
@@ -76,7 +76,7 @@ namespace SFA.DAS.Commitments.Shared.Services
             return _client.UpdateDraftApprenticeship(cohortId, draftApprenticeshipId, updateRequest);
         }
 
-        public async Task<GetApprenticeshipsResponse> GetApprenticeships(uint providerId, string sortField = "", bool reverseSort = false)
+        public async Task<GetApprenticeshipsResponse> GetApprenticeships(long providerId, string sortField = "", bool reverseSort = false)
         {
             var getApprenticeshipsResponse = await _client.GetApprenticeships(new GetApprenticeshipRequest{
 			ProviderId = providerId,
